--- access_control/grant_owner/GrantOwnerGateway.go
@@ -1,9 +1,0 @@
-package grant_owner
-
-import "github.com/attestify/go-kernel/access_control"
-
-type GrantOwnerGateway interface {
-	Grant(accessControl access_control.AccessControl)
-	HasError() bool
-	Error() error
-}
