+++ authorization/modify_access/ModifyAccessGateway.go
@@ -1,6 +1,6 @@
 package modify_access
 
-import "github.com/attestify/go-kernel/access_control"
+import "github.com/attestify/go-kernel/authorization"
 
 // ModifyAccessGateway provides the behavior for persisting a role record
 // Expected Alternative Behaviors
@@ -9,7 +9,7 @@ import "github.com/attestify/go-kernel/access_control"
 // Returns one (1) type of error:
 //  - InternalError - Returned if there is an error invoking the implementation of this interface
 type ModifyAccessGateway interface {
-	Modify(accessControl access_control.AccessControl)
+	Modify(accessControl authorization.AccessControl)
 	Error() error
 	HasError() bool
 }
