+++ authorization/revoke_access/RevokeAccess_test.go
@@ -2,8 +2,8 @@ package revoke_access_test
 
 import (
 	"errors"
-	"github.com/attestify/go-kernel/access_control"
-	"github.com/attestify/go-kernel/access_control/revoke_access"
+	"github.com/attestify/go-kernel/authorization"
+	"github.com/attestify/go-kernel/authorization/revoke_access"
 	"github.com/attestify/go-kernel/error/internal_error"
 	"testing"
 )
@@ -188,7 +188,7 @@ func NewMockRevokeAccessGateway() MockRevokeAccessGateway {
 	return MockRevokeAccessGateway{}
 }
 
-func (gateway *MockRevokeAccessGateway) Revoke(accessControl access_control.AccessControl) {
+func (gateway *MockRevokeAccessGateway) Revoke(accessControl authorization.AccessControl) {
 	if gateway.revokeAccessGatewayInternalError {
 		gateway.gatewayError = internal_error.New("Error generated from RevokeAccess.Revoke() invocation.")
 	}
